+++ website/src/routes/(docs)/qwik/api/(types)/FieldStore/properties.ts
@@ -125,6 +125,14 @@ export const properties: Record<string, PropertyProps> = {
           },
         },
         {
+          key: 'isEdited',
+          value: {
+            type: 'custom',
+            name: 'ReadonlySignal',
+            generics: ['boolean'],
+          },
+        },
+        {
           key: 'isDirty',
           value: {
             type: 'custom',
@@ -291,6 +299,13 @@ export const properties: Record<string, PropertyProps> = {
     },
   },
   isTouched: {
+    type: {
+      type: 'custom',
+      name: 'ReadonlySignal',
+      generics: ['boolean'],
+    },
+  },
+  isEdited: {
     type: {
       type: 'custom',
       name: 'ReadonlySignal',